Allow draw updates while paused
Previously when the simulation was paused, the draw call returned immediately. This prevented updates to the grid via user input from showing up until the simulation was started again. Now the ants and grid are drawn every frame, regardless of whether the simulation is paused; only updates to the ants and grid (movement; pheromones; etc) are skipped while paused.